changed around the add and remove favorite to the emoji in the feedcard instead of an option in the modal
