+++ src/components/dashboard/RecentCard.js
@@ -13,7 +13,7 @@ import Avatar from "@material-ui/core/Avatar";
 import IconButton from "@material-ui/core/IconButton";
 import Typography from "@material-ui/core/Typography";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 // import ChatBubbleIcon from "@material-ui/icons/ChatBubble";
 // import { Icon } from "@iconify/react";
 // import smHeart from "@iconify/icons-heroicons/sm-heart";
@@ -30,12 +30,23 @@ import { axiosWithAuth } from "../../utilities/axiosWithAuth";
 import { USER_BY_ID } from "../../graphql/users/user-queries";
 import { print } from "graphql";
 
+import {
+  ADD_FAVORITE_EVENT,
+  REMOVE_FAVORITE_EVENT,
+} from "../../graphql/users/user-mutations";
+
+import {
+  addFavoriteEventSuccess,
+  removeFavoriteEventSuccess,
+} from "../../utilities/actions";
+
 import EventButtonModal from "./EventButtonModal";
 import Emoji from "../other/Emoji";
 
 const RecentCard = (props) => {
   const me = JSON.parse(sessionStorage.getItem("user"));
   const classes = cardStyles();
+  const dispatch = useDispatch();
   const [expanded, setExpanded] = useState(false);
   const [currentStatus, setCurrentStatus] = useState(props.currentStatus);
   const [creatorName, setCreatorName] = useState("");
@@ -54,6 +65,39 @@ const RecentCard = (props) => {
   // const toggleLike = () => {
   //   setLiked(!liked);
   // };
+  const addFavoriteEvent = () => {
+    const addFavorite = {
+      event_id: Number(props.id),
+      user_id: Number(me.id),
+    };
+
+    axiosWithAuth()
+      .post(`${process.env.REACT_APP_BASE_URL}/graphql`, {
+        query: print(ADD_FAVORITE_EVENT),
+        variables: { input: addFavorite },
+      })
+      .then((res) => {
+        dispatch(addFavoriteEventSuccess(res.data.data.addFavoriteEvent));
+      })
+      .catch((err) => console.log(err));
+  };
+
+  const removeFavoriteEvent = () => {
+    const removeFavorite = {
+      event_id: Number(props.id),
+      user_id: Number(me.id),
+    };
+
+    axiosWithAuth()
+      .post(`${process.env.REACT_APP_BASE_URL}/graphql`, {
+        query: print(REMOVE_FAVORITE_EVENT),
+        variables: { input: removeFavorite },
+      })
+      .then((res) => {
+        dispatch(removeFavoriteEventSuccess(res.data.data.removeFavoriteEvent));
+      })
+      .catch((err) => console.log(err));
+  };
 
   useEffect(() => {
     //get creator name when event loads.  This is a rough and inefficient way to do this, especially if there ends up being protected queries
@@ -154,8 +198,18 @@ const RecentCard = (props) => {
         {/* <Typography variant="caption" color="textSecondary">
           RSVP
         </Typography> */}
-        {isFavorite && (
-          <div style={{ fontSize: "2.5rem" }}>
+        {!isFavorite ? (
+          <div
+            style={{ fontSize: "2.5rem", cursor: "pointer" }}
+            onClick={() => addFavoriteEvent()}
+          >
+            <Emoji label="star" symbol="☆" />
+          </div>
+        ) : (
+          <div
+            style={{ fontSize: "2.5rem", cursor: "pointer" }}
+            onClick={() => removeFavoriteEvent()}
+          >
             <Emoji label="star" symbol="⭐" />
           </div>
         )}
